Rename effects for hooks
